feat: renew function

+++ app/Services/CartService.php
@@ -23,29 +23,43 @@ class CartService
         return $user->cart()->delete();
     }
 
-    public function add($request): bool
+    public function add($product_id, $cycle_id, $amount = 1, $renew_for = null): bool
     {
         $cart = $this->getUserCart($this->user);
-        $product_in_cart = $cart->where('product_id', $request->product_id)->where('cycle_id', $request->cycle_id)->first();
-        if ($product_in_cart) {
-            $product_in_cart->amount += $request->amount;
-            $product_in_cart->save();
-            return true;
+
+        if ($renew_for) {
+            $product_in_cart = $cart->where('product_id', $product_id)->where('renew_for', $renew_for)->first();
+            if ($product_in_cart) {
+                return true;
+            }
+        } else {
+            $product_in_cart = $cart->where('product_id', $product_id)->where('cycle_id', $cycle_id)->whereNull('renew_for')->first();
+            if ($product_in_cart) {
+                $product_in_cart->amount += $amount;
+                $product_in_cart->save();
+                return true;
+            }
         }
 
-        $product = Product::where('id', $request->product_id)->where('status', 1)->first();
-        $cycle = $product->productCycle()->where('id', $request->cycle_id)->first();
-        $amount = $request->amount;
+        $product = Product::where('id', $product_id)->where('status', 1)->first();
+        if (!$product) {
+            return false;
+        }
+        $cycle = $product->productCycle()->where('id', $cycle_id)->first();
         $this->user->cart()->create([
             'product_id' => $product->id,
             'cycle_id' => $cycle->id,
-            'amount' => $amount
+            'amount' => $amount,
+            'renew_for' => $renew_for ?? null
         ]);
         return true;
     }
     public function remove($request): bool
     {
         $cart_item = $this->user->cart()->where('id', $request->cart_item_id)->first();
+        if (!$cart_item) {
+            return false;
+        }
         $cart_item->delete();
         return true;
     }
@@ -53,6 +67,12 @@ class CartService
     public function update($request): bool
     {
         $cart_item = $this->user->cart()->where('id', $request->cart_item_id)->first();
+        if (!$cart_item) {
+            return false;
+        }
+        if ($cart_item->renew_for) {
+            return false;
+        }
         // return $request;
         $cart_item->amount = $request->amount;
         $cart_item->save();
